minijinja::functions

Function debug

pub fn debug(state: &State<'_, '_>, args: Rest<Value>) -> String

Outputs the current context or the arguments stringified.

This is a useful function to quickly figure out the state of affairs in a template. It emits a stringified debug dump of the current engine state including the layers of the context, the current block and auto escaping setting. The exact output is not defined and might change from one version of Jinja2 to the next.

<pre>{{ debug() }}</pre>
<pre>{{ debug(variable1, variable2) }}</pre>