Struct minijinja::State

pub struct State<'template, 'env> { /* private fields */ }

Provides access to the current execution state of the engine.

A read only reference is passed to filter functions and similar objects to allow limited interfacing with the engine. The state is useful to look up information about the engine in filter, test or global functions. It not only provides access to the template environment but also the context variables of the engine, the current auto escaping behavior as well as the auto escape flag.

In some testing scenarios or more advanced use cases you might need to get a State. The state is managed as part of the template execution but the initial state can be retrieved via Template::new_state. The most common way to get hold of the state however is via functions of filters.

Notes on lifetimes: the state object exposes some of the internal lifetimes through the type. You should always elide these lifetimes as there might be lifetimes added or removed between releases.

Implementations

impl<'template, 'env> State<'template, 'env>

pub fn env(&self) -> &Environment<'_>

Returns a reference to the current environment.

pub fn name(&self) -> &str

Returns the name of the current template.

pub fn auto_escape(&self) -> AutoEscape

Returns the current value of the auto escape flag.

pub fn undefined_behavior(&self) -> UndefinedBehavior

Returns the current undefined behavior.

pub fn current_block(&self) -> Option<&str>

Returns the name of the innermost block.

pub fn lookup(&self, name: &str) -> Option<Value>

Looks up a variable by name in the context.

pub fn call_macro(&self, name: &str, args: &[Value]) -> Result<String, Error>

Looks up a global macro and calls it.

This looks up a value as lookup does and calls it with the passed args.

pub fn render_block(&mut self, block: &str) -> Result<String, Error>

Renders a block with the given name into a string.

This method works like Template::render but it only renders a specific block in the template. The first argument is the name of the block.

This renders only the block hi in the template:

let tmpl = env.get_template("hello")?;
let rv = tmpl
    .eval_to_state(context!(name => "John"))?
    .render_block("hi")?;
println!("{}", rv);

Note that rendering a block is a stateful operation. If an error is returned the module has to be re-created as the internal state can end up corrupted. This also means you can only render blocks if you have a mutable reference to the state which is not possible from within filters or similar.

pub fn render_block_to_write<W>( &mut self, block: &str, w: W ) -> Result<(), Error>
where W: Write,

Renders a block with the given name into an io::Write.

For details see render_block.

pub fn exports(&self) -> Vec<&str>

Returns a list of the names of all exports (top-level variables).

pub fn get_template(&self, name: &str) -> Result<Template<'env, 'env>, Error>

Fetches a template by name with path joining.

This works like Environment::get_template with the difference that the lookup undergoes path joining. If the environment has a configured path joining callback, it will be invoked with the name of the current template as parent template.

For more information see Environment::set_path_join_callback.

pub fn apply_filter(&self, filter: &str, args: &[Value]) -> Result<Value, Error>

Invokes a filter with some arguments.

let rv = state.apply_filter("upper", &["hello world".into()]).unwrap();
assert_eq!(rv.as_str(), Some("HELLO WORLD"));

pub fn perform_test(&self, test: &str, args: &[Value]) -> Result<bool, Error>

Invokes a test function on a value.

let rv = state.perform_test("even", &[42i32.into()]).unwrap();
assert!(rv);

pub fn format(&self, value: Value) -> Result<String, Error>

Formats a value to a string using the formatter on the environment.

let rv = state.format(Value::from(42)).unwrap();
assert_eq!(rv, "42");