Function minijinja::value::from_args

pub fn from_args<'a, Args>(values: &'a [Value]) -> Result<Args, Error> where
    Args: FunctionArgs<'a, Output = Args>, 

Utility function to convert a slice of values into arguments.

This performs the same conversion that Function performs. It exists so that you one can leverage the same functionality when implementing Object::call_method.

use minijinja::value::from_args;

// args is &[Value]
let (string, num): (&str, i64) = from_args(args)?;

Note that only value conversions are supported which means that &State is not a valid conversion type.