Struct minijinja::value::DynObject

pub struct DynObject { /* private fields */ }

Typed-erased version of Object

Implementations

impl DynObject

pub fn new<T: Object + Send + Sync + 'static>(v: Arc<T>) -> Self

Returns a new boxed, type-erased Object.

pub fn repr(&self) -> ObjectRepr

Calls Object::repr on the underlying boxed value.

pub fn get_value(&self, key: &Value) -> Option<Value>

Calls Object::get_value on the underlying boxed value.

pub fn enumerate(&self) -> Enumerator

Calls Object::enumerate on the underlying boxed value.

pub fn is_true(&self) -> bool

Calls Object::is_true on the underlying boxed value.

pub fn enumerator_len(&self) -> Option<usize>

Calls Object::enumerator_len on the underlying boxed value.

pub fn call( &self, state: &State<'_, '_>, args: &[Value] ) -> Result<Value, Error>

Calls Object::call on the underlying boxed value.

pub fn call_method( &self, state: &State<'_, '_>, method: &str, args: &[Value] ) -> Result<Value, Error>

Calls Object::call_method on the underlying boxed value.

pub fn render(&self, f: &mut Formatter<'_>) -> Result

Calls Object::render on the underlying boxed value.

pub fn type_name(&self) -> &'static str

Returns the type name of the conrete underlying type.

pub fn downcast_ref<T: 'static>(&self) -> Option<&T>

Downcast to T if the boxed value holds a T.

This is basically the “reverse” of Value::from_object.

Example

use std::fmt;

#[derive(Debug)]
struct Thing {
    id: usize,
}

impl Object for Thing {}

let x_value = Value::from_object(Thing { id: 42 });
let value_as_obj = x_value.as_object().unwrap();
let thing = value_as_obj.downcast_ref::<Thing>().unwrap();
assert_eq!(thing.id, 42);

pub fn downcast<T: 'static>(&self) -> Option<Arc<T>>

Downcast to T if the boxed value holds a T.

For details see downcast_ref.

pub fn is<T: 'static>(&self) -> bool

Checks if the boxed value is a T.

For details see downcast_ref.

impl DynObject

pub fn try_iter(&self) -> Option<Box<dyn Iterator<Item = Value> + Send + Sync>>
where Self: 'static,

Iterates over this object.

If this returns None then the default object iteration as defined by the object’s enumeration is used.

pub fn try_iter_pairs( &self ) -> Option<Box<dyn Iterator<Item = (Value, Value)> + Send + Sync>>

Iterate over key and value at once.