Struct minijinja::value::Value

pub struct Value(_);

Represents a dynamically typed value in the template engine.

Implementations

impl Value

pub const UNDEFINED: Value = _

The undefined value

pub fn from_serializable<T: Serialize>(value: &T) -> Value

Creates a value from something that can be serialized.

During serialization of the value, serializing_for_value will return true which makes it possible to customize serialization for MiniJinja. For more information see serializing_for_value.

pub fn from_safe_string(value: String) -> Value

Creates a value from a safe string.

pub fn from_object<T: Object + 'static>(value: T) -> Value

Creates a value from a dynamic object.

pub fn from_function<F, Rv, Args>(f: F) -> Value where
F: Function<Rv, Args> + for<'a> Function<Rv, <Args as FunctionArgs<'a>>::Output>,
Rv: FunctionResult,
Args: for<'a> FunctionArgs<'a>,

Creates a callable value from a function.

pub fn downcast_object_ref<T: Object>(&self) -> Option<&T>

Returns some reference to the boxed object if it is of type T, or None if it isn’t.

This is basically the “reverse” of from_object.

Example

use std::fmt;

#[derive(Debug)]
struct Thing {
    id: usize,
}

impl fmt::Display for Thing {
    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
        fmt::Debug::fmt(self, f)
    }
}

impl Object for Thing {}

let x_value = Value::from_object(Thing { id: 42 });
let thing = x_value.downcast_object_ref::<Thing>().unwrap();
assert_eq!(thing.id, 42);

pub fn kind(&self) -> ValueKind

Returns the value kind.

pub fn is_kwargs(&self) -> bool

Returns true if the map represents keyword arguments.

pub fn as_str(&self) -> Option<&str>

If the value is a string, return it.

pub fn as_bytes(&self) -> Option<&[u8]>

Returns the bytes of this value if they exist.

pub fn is_true(&self) -> bool

Is this value true?

pub fn is_safe(&self) -> bool

Returns true if this value is safe.

pub fn is_undefined(&self) -> bool

Returns true if this value is undefined.

pub fn is_none(&self) -> bool

Returns true if this value is none.

pub fn len(&self) -> Option<usize>

Returns the length of the contained value.

pub fn get_attr(&self, key: &str) -> Result<Value, Error>

Looks up an attribute by attribute name.

This this returns UNDEFINED when an invalid key is resolved. An error is returned when if the value does not contain an object that has attributes.

pub fn get_item_by_index(&self, idx: usize) -> Result<Value, Error>

Looks up an index of the value.

This is a shortcut for get_item.

pub fn get_item(&self, key: &Value) -> Result<Value, Error>

Looks up an item (or attribute) by key.

This is similar to get_attr but instead of using a string key this can be any key. For instance this can be used to index into sequences. Like get_attr this returns UNDEFINED when an invalid key is looked up.

pub fn as_slice(&self) -> Result<&[Value], Error>

If the value is a sequence it’s returned as slice.