Module minijinja::tests

Test functions and abstractions.

Test functions in MiniJinja are like filters but a different syntax is used to invoke them and they have to return boolean values. For instance the expression {% if foo is defined %} invokes the is_defined test to check if the value is indeed an odd number.

MiniJinja comes with some built-in test functions that are listed below. To create a custom test write a function that takes at least a value argument that returns a boolean result, then register it with add_filter.

Using Tests

Tests are useful to “test” a value in a specific way. For instance if you want to assign different classes to alternating rows one way is using the odd test:

{% if seq is defined %}
  <ul>
  {% for item in seq %}
    <li class="{{ 'even' if loop.index is even else 'odd' }}">{{ item }}</li>
  {% endfor %}
  </ul>
{% endif %}

Custom Tests

A custom test function is just a simple function which accepts its inputs as parameters and then returns a bool. For instance the following shows a test function which takes an input value and checks if it’s lowercase:

fn is_lowercase(value: String) -> bool {
    value.chars().all(|x| x.is_lowercase())
}

env.add_test("lowercase", is_lowercase);

MiniJinja will perform the necessary conversions automatically. For more information see the Test trait.

Built-in Tests

When the builtins feature is enabled a range of built-in tests are automatically added to the environment. These are also all provided in this module. Note though that these functions are not to be called from Rust code as their exact interface (arguments and return types) might change from one MiniJinja version to another.

Traits

Test

A utility trait that represents test functions.

TestResult

A utility trait that represents the return value of filters.

Functions

is_defined

Checks if a value is defined.

is_endingwith

Checks if the value is ending with a string.

is_eq

Test version of ==.

is_even

Checks if a value is even.

is_ge

Test version of >=.

is_gt

Test version of >.

is_in

Test version of in.

is_le

Test version of <=.

is_lt

Test version of <.

is_mapping

Checks if this value is a mapping

is_ne

Test version of !=.

is_none

Checks if a value is none.

is_number

Checks if this value is a number.

is_odd

Checks if a value is odd.

is_safe

Checks if a value is safe.

is_sequence

Checks if this value is a sequence

is_startingwith

Checks if the value is starting with a string.

is_string

Checks if this value is a string.

is_undefined

Checks if a value is undefined.