minijinja::value

Struct DynObject

pub struct DynObject { /* private fields */ }

Type-erased version of Object

Implementations

impl DynObject

pub fn new<T: Object + 'static>(v: Arc<T>) -> Self

Returns a new boxed, type-erased Object.

pub fn repr(&self) -> ObjectRepr

Calls Object::repr of the underlying boxed value.

pub fn get_value(&self, key: &Value) -> Option<Value>

Calls Object::get_value of the underlying boxed value.

pub fn enumerate(&self) -> Enumerator

Calls Object::enumerate of the underlying boxed value.

pub fn is_true(&self) -> bool

Calls Object::is_true of the underlying boxed value.

pub fn enumerator_len(&self) -> Option<usize>

Calls Object::enumerator_len of the underlying boxed value.

pub fn call( &self, state: &State<'_, '_>, args: &[Value], ) -> Result<Value, Error>

Calls Object::call of the underlying boxed value.

pub fn call_method( &self, state: &State<'_, '_>, method: &str, args: &[Value], ) -> Result<Value, Error>

Calls Object::call_method of the underlying boxed value.

pub fn custom_cmp(&self, other: &DynObject) -> Option<Ordering>

Calls Object::custom_cmp of the underlying boxed value.

pub fn render(&self, f: &mut Formatter<'_>) -> Result

Calls Object::render of the underlying boxed value.

pub fn type_name(&self) -> &'static str

Returns the type name of the concrete underlying type.

pub fn downcast_ref<T: 'static>(&self) -> Option<&T>

Downcast to T if the boxed value holds a T.

This works like Any::downcast_ref.

pub fn downcast<T: 'static>(&self) -> Option<Arc<T>>

Downcast to T if the boxed value holds a T.

This is similar to downcast_ref but returns the Arc.

pub fn is<T: 'static>(&self) -> bool

Checks if the boxed value is a T.

This works like Any::is.

impl DynObject

pub fn try_iter(&self) -> Option<Box<dyn Iterator<Item = Value> + Send + Sync>>
where Self: 'static,

Iterates over this object.

If this returns None then the default object iteration as defined by the objectโs enumeration is used.

pub fn try_iter_pairs( &self, ) -> Option<Box<dyn Iterator<Item = (Value, Value)> + Send + Sync>>

Iterate over key and value at once.