Struct minijinja::value::Value

pub struct Value(/* private fields */);

Represents a dynamically typed value in the template engine.

Implementations

impl Value

pub const UNDEFINED: Value = _

The undefined value.

This constant exists because the undefined type does not exist in Rust and this is the only way to construct it.

pub fn from_serialize<T: Serialize>(value: T) -> Value

Creates a value from something that can be serialized.

This is the method that MiniJinja will generally use whenever a serializable object is passed to one of the APIs that internally want to create a value. For instance this is what context! and render will use.

During serialization of the value, serializing_for_value will return true which makes it possible to customize serialization for MiniJinja. For more information see serializing_for_value.

let val = Value::from_serialize(&vec![1, 2, 3]);

This method does not fail but it might return a value that is not valid. Such values will when operated on fail in the template engine in most situations. This for instance can happen if the underlying implementation of Serialize fails. There are also cases where invalid objects are silently hidden in the engine today. This is for instance the case for when keys are used in hash maps that the engine cannot deal with. Invalid values are considered an implementation detail. There is currently no API to validate a value.

If the deserialization feature is enabled then the inverse of this method is to use the Value type as serializer. You can pass a value into the deserialize method of a type that supports serde deserialization.

pub fn from_safe_string(value: String) -> Value

Creates a value from a safe string.

A safe string is one that will bypass auto escaping. For instance if you want to have the template engine render some HTML without the user having to supply the |safe filter, you can use a value of this type instead.

let val = Value::from_safe_string("<em>note</em>".into());

pub fn from_object<T: Object + Send + Sync + 'static>(value: T) -> Value

Creates a value from a dynamic object.

For more information see Object.

use std::fmt;

#[derive(Debug)]
struct Thing {
    id: usize,
}

impl Object for Thing {}

let val = Value::from_object(Thing { id: 42 });

pub fn from_dyn_object<T: Into<DynObject>>(value: T) -> Value

Like from_object but for type erased dynamic objects.

This especially useful if you have an object that has an Arc<T> to another child object that you want to return as a Arc<T> turns into a DynObject automatically.

#[derive(Debug)]
pub struct HttpConfig {
    port: usize,
}

#[derive(Debug)]
struct Config {
    http: Arc<HttpConfig>,
}

impl Object for HttpConfig {
    fn enumerate(self: &Arc<Self>) -> Enumerator {
        Enumerator::Str(&["port"])
    }

    fn get_value(self: &Arc<Self>, key: &Value) -> Option<Value> {
        match key.as_str()? {
            "port" => Some(Value::from(self.port)),
            _ => None,
        }
    }
}

impl Object for Config {
    fn enumerate(self: &Arc<Self>) -> Enumerator {
        Enumerator::Str(&["http"])
    }

    fn get_value(self: &Arc<Self>, key: &Value) -> Option<Value> {
        match key.as_str()? {
            "http" => Some(Value::from_dyn_object(self.http.clone())),
            _ => None
        }
    }
}

pub fn make_iterable<I, T, F>(maker: F) -> Value
where I: Iterator<Item = T> + Send + Sync + 'static, T: Into<Value> + Send + Sync + 'static, F: Fn() -> I + Send + Sync + 'static,

Creates a value that is an iterable.

The function is invoked to create a new iterator every time the value is iterated over.

let val = Value::make_iterable(|| 0..10);

Iterators that implement ExactSizeIterator or have a matching lower and upper bound on the Iterator::size_hint report a known loop.length. Iterators that do not fulfill these requirements will not. The same is true for revindex and similar properties.

pub fn make_object_iterable<T, F>(object: T, maker: F) -> Value
where T: Send + Sync + 'static, F: for<'a> Fn(&'a T) -> Box<dyn Iterator<Item = Value> + Send + Sync + 'a> + Send + Sync + 'static,

Creates an iterable that iterates over the given value.

This is similar to make_iterable but it takes an extra reference to a value it can borrow out from. It’s a bit less generic in that it needs to return a boxed iterator of values directly.

let val = Value::make_object_iterable(vec![1, 2, 3], |vec| {
    Box::new(vec.iter().copied().map(Value::from))
});
assert_eq!(val.to_string(), "[1, 2, 3]");

pub fn make_one_shot_iterator<I, T>(iter: I) -> Value
where I: Iterator<Item = T> + Send + Sync + 'static, T: Into<Value> + Send + Sync + 'static,

Creates a value from a one-shot iterator.

This takes an iterator (yielding values that can be turned into a Value) and wraps it in a way that it turns into an iterable value. From the view of the template this can be iterated over exactly once for the most part once exhausted.

Such iterators are strongly recommended against in the general sense due to their surprising behavior, but they can be useful for more advanced use cases where data should be streamed into the template as it becomes available.

Such iterators never have any size hints.

let val = Value::make_one_shot_iterator(0..10);

Attempting to iterate over it a second time will not yield any more items.

pub fn from_function<F, Rv, Args>(f: F) -> Value
where F: Function<Rv, Args> + for<'a> Function<Rv, <Args as FunctionArgs<'a>>::Output>, Rv: FunctionResult, Args: for<'a> FunctionArgs<'a>,

Creates a callable value from a function.

let pow = Value::from_function(|a: u32| a * a);

pub fn kind(&self) -> ValueKind

Returns the kind of the value.

This can be used to determine what’s in the value before trying to perform operations on it.

pub fn is_number(&self) -> bool

Returns true if the value is a number.

To convert a value into a primitive number, use TryFrom or TryInto.

pub fn is_kwargs(&self) -> bool

Returns true if the map represents keyword arguments.

pub fn is_true(&self) -> bool

Is this value considered true?

The engine inherits the same behavior as Jinja2 when it comes to considering objects true. Empty objects are generally not considered true. For custom objects this is customized by Object::is_true.

pub fn is_safe(&self) -> bool

Returns true if this value is safe.

pub fn is_undefined(&self) -> bool

Returns true if this value is undefined.

pub fn is_none(&self) -> bool

Returns true if this value is none.

pub fn to_str(&self) -> Option<Arc<str>>

If the value is a string, return it.

pub fn as_str(&self) -> Option<&str>

If the value is a string, return it.

pub fn as_usize(&self) -> Option<usize>

If this is an i64 return it

pub fn as_i64(&self) -> Option<i64>

If this is an i64 return it

pub fn as_bytes(&self) -> Option<&[u8]>

Returns the bytes of this value if they exist.

pub fn as_object(&self) -> Option<&DynObject>

If the value is an object a reference to it is returned.

The returned value is a reference to a type erased DynObject. For a specific type use downcast_object instead.

pub fn len(&self) -> Option<usize>

Returns the length of the contained value.

Values without a length will return None.

let seq = Value::from(vec![1, 2, 3, 4]);
assert_eq!(seq.len(), Some(4));

pub fn get_attr(&self, key: &str) -> Result<Value, Error>

Looks up an attribute by attribute name.

This this returns UNDEFINED when an invalid key is resolved. An error is returned if the value does not contain an object that has attributes.

let ctx = minijinja::context! {
    foo => "Foo"
};
let value = ctx.get_attr("foo")?;
assert_eq!(value.to_string(), "Foo");

pub fn get_item_by_index(&self, idx: usize) -> Result<Value, Error>

Looks up an index of the value.

This is a shortcut for get_item.

let seq = Value::from(vec![0u32, 1, 2]);
let value = seq.get_item_by_index(1).unwrap();
assert_eq!(value.try_into().ok(), Some(1));

pub fn get_item(&self, key: &Value) -> Result<Value, Error>

Looks up an item (or attribute) by key.

This is similar to get_attr but instead of using a string key this can be any key. For instance this can be used to index into sequences. Like get_attr this returns UNDEFINED when an invalid key is looked up.

let ctx = minijinja::context! {
    foo => "Foo",
};
let value = ctx.get_item(&Value::from("foo")).unwrap();
assert_eq!(value.to_string(), "Foo");

pub fn try_iter(&self) -> Result<ValueIter, Error>

Iterates over the value.

Depending on the kind of the value the iterator has a different behavior.

ValueKind::Map: the iterator yields the keys of the map.

ValueKind::Seq / ValueKind::Iterable: the iterator yields the items in the sequence.

ValueKind::String: the iterator yields characters in a string.

ValueKind::None / ValueKind::Undefined: the iterator is empty.

let value = Value::from({
    let mut m = std::collections::BTreeMap::new();
    m.insert("foo", 42);
    m.insert("bar", 23);
    m
});
for key in value.try_iter()? {
    let value = value.get_item(&key)?;
    println!("{} = {}", key, value);
}

pub fn reverse(&self) -> Result<Value, Error>

Returns a reversed view of this value.

This is implemented for the following types with the following behaviors:

undefined or none: value returned unchanged.

string and bytes: returns a reversed version of that value

iterables: returns a reversed version of the iterable. If the iterable is not reversable itself, it consumes it and then reverses it.

pub fn downcast_object_ref<T: 'static>(&self) -> Option<&T>

Returns some reference to the boxed object if it is of type T, or None if it isn’t.

This is basically the “reverse” of from_object, from_object and from_object. It’s also a shortcut for downcast_ref on the return value of as_object.

Example

use std::fmt;

#[derive(Debug)]
struct Thing {
    id: usize,
}

impl Object for Thing {}

let x_value = Value::from_object(Thing { id: 42 });
let thing = x_value.downcast_object_ref::<Thing>().unwrap();
assert_eq!(thing.id, 42);

pub fn downcast_object<T: 'static>(&self) -> Option<Arc<T>>

Like downcast_object_ref but returns the actual object.

pub fn call( &self, state: &State<'_, '_>, args: &[Value] ) -> Result<Value, Error>

Calls the value directly.

If the value holds a function or macro, this invokes it. Note that in MiniJinja there is a separate namespace for methods on objects and callable items. To call methods (which should be a rather rare occurrence) you have to use call_method.

The args slice is for the arguments of the function call. To pass keyword arguments use the Kwargs type.

Usually the state is already available when it’s useful to call this method, but when it’s not available you can get a fresh template state straight from the Template via new_state.

let func = Value::from_function(|v: i64, kwargs: Kwargs| {
    v * kwargs.get::<i64>("mult").unwrap_or(1)
});
let rv = func.call(
    state,
    &[
        Value::from(42),
        Value::from(Kwargs::from_iter([("mult", Value::from(2))])),
    ],
).unwrap();
assert_eq!(rv, Value::from(84));

With the args! macro creating an argument slice is simplified:

let func = Value::from_function(|v: i64, kwargs: Kwargs| {
    v * kwargs.get::<i64>("mult").unwrap_or(1)
});
let rv = func.call(state, args!(42, mult => 2)).unwrap();
assert_eq!(rv, Value::from(84));

pub fn call_method( &self, state: &State<'_, '_>, name: &str, args: &[Value] ) -> Result<Value, Error>

Calls a method on the value.

The name of the method is name, the arguments passed are in the args slice.