Trait minijinja::value::FunctionArgs

pub trait FunctionArgs<'a> {
    type Output;
}

Helper trait representing valid filter, test and function arguments.

Since it’s more convenient to write filters and tests with concrete types instead of values, this helper trait exists to automatically perform this conversion. It is implemented for functions up to an arity of 5 parameters.

For each argument the conversion is performed via the ArgType trait which is implemented for many common types. For manual conversions the from_args utility should be used.

Required Associated Types

type Output

The output type of the function arguments.

Object Safety

This trait is not object safe.